+++ tests/test_simulator.py
@@ -1,8 +1,7 @@
-import jax 
-import jax.numpy as jnp 
-import pandas as pd 
+import jax
+import jax.numpy as jnp
 
-from dynax.simulators.simulator import DifferentiableSimulator
+from simulator import DifferentiableSimulator
 from dynax.models.RC import Continuous4R3C
 from dynax.agents import Tabular
 
@@ -16,15 +15,14 @@ output_dim = model.output_dim
 # resample to a given time step
 dt = 1.
 
-
 # forward simulation settings
 ts = 0.
 te = 10.
 n_steps = int((te - ts) / dt)
 x0 = jnp.array([20., 30., 26.]) # initial state
 
-RESULTS = jnp.array(
-    [[ 2.0000e+01,3.0000e+01,2.6000e+01],
+RESULTS = jnp.array([
+    [ 2.0000e+01,3.0000e+01,2.6000e+01],
     [ 9.0000e+00,-2.0000e+00,2.5000e+01],
     [ 1.9000e+01,2.9000e+01,-1.7000e+01],
     [-3.3000e+01,-4.4000e+01,6.6000e+01],
@@ -37,44 +35,27 @@ RESULTS = jnp.array(
     [ 2.0479e+04,2.0489e+04,-2.8961e+04]]
 )
 
-def test_tabular_inputs():
+def test_open_loop():
+    # inputs
+    dist = jnp.ones((n_steps,4))
+    policy = jnp.ones((n_steps,1))
 
-    # instantiate a model
+    # gather for RC model inputs
+    inputs = jnp.concatenate([dist[:,:2], policy, dist[:, 2:]], axis=1)
+
+    # simulator
     simulator = DifferentiableSimulator(
         model = model,
         dt = dt
     )
 
-    # given input sequences
-    inputs = jnp.ones((n_steps+1,5))
-    inputs_tab = Tabular(ts = jnp.arange(ts, te+dt, dt), xs=inputs)
-
     # need initialize the simulator first
-    inits = simulator.init(jax.random.PRNGKey(0), inputs_tab, x0, ts, te)
+    inits = simulator.init(jax.random.PRNGKey(0), x0, inputs)
 
     # simulate forward problem
-    _, xsol, ysol = simulator.apply(inits, inputs_tab, x0, ts, te)
-    #_, xsol, ysol = simulator(x0)
-    
-    assert jnp.allclose(xsol, RESULTS), "tabular agent didn't get expected results."
+    xsol, ysol = simulator.apply(inits, x0, inputs)
+    assert jnp.allclose(xsol, RESULTS[1:,:]), "closed loop test didn't get expected results."
 
-
-def test_array_inputs():
-
-    simulator = DifferentiableSimulator(
-    #    state = states,
-        model = model,
-        dt = dt
-    )
-    inputs = jnp.array([1., 1., 1., 1., 1.])
-
-    # need initialize the simulator first
-    inits = simulator.init(jax.random.PRNGKey(0), inputs, x0, ts, te)
-
-    # simulate forward problem
-    _, xsol, ysol = simulator.apply(inits, inputs, x0, ts, te)
-
-    assert jnp.allclose(xsol, RESULTS), "tabular agent didn't get expected results."
 
 def test_closed_loop():
 
@@ -86,27 +67,23 @@ def test_closed_loop():
     model = model,
     dt = dt,
     )
-    inits = simulator.init(jax.random.PRNGKey(0), jnp.array([1., 1., 1., 1., 1.]), x_prev, ts, te)
+    inits = simulator.init(jax.random.PRNGKey(0), x_prev, jnp.ones((1, model.input_dim)))
 
     while t < te:
         dist = jnp.array([1., 1., 1., 1.])
         policy = 1.
-        inputs = jnp.concatenate([dist, jnp.array(policy).reshape(-1)])
+        inputs = jnp.concatenate([dist, jnp.array(policy).reshape(-1)]).reshape(1,-1)
 
-        _, xsol, ysol = simulator.apply(inits, inputs, x_prev, t, t+dt)
+        xsol, ysol = simulator.apply(inits, x_prev, inputs)
 
         x_prev = xsol[-1,:]
         res.append(x_prev)
         t += dt 
     
     res = jnp.stack(res, axis=0)
-
     assert jnp.allclose(res, RESULTS), "closed loop test didn't get expected results."
-    return res
 
 
 if __name__ == "__main__":
-    test_tabular_inputs()
-    test_array_inputs()
+    test_open_loop()
     test_closed_loop()
-    print("all test in forward simulator passed !!!!")
